clean up big component
